move initApp into actions
