+++ client/src/stores/queries-actions.ts
@@ -10,6 +10,10 @@ import {
 import runQueryViaBatch from '../utilities/runQueryViaBatch';
 import { NEW_QUERY, useQueriesStore } from './queries-store';
 import localforage from 'localforage';
+import queryString from 'query-string';
+
+// @ts-expect-error
+window.localforage = localforage;
 
 export function useSelectedConnectionId(): string {
   return useQueriesStore((s) => s.selectedConnectionId);
@@ -18,6 +22,65 @@ export function useSelectedConnectionId(): string {
 export function useConnectionClient(): any {
   return useQueriesStore((s) => s.connectionClient);
 }
+
+export const initApp = async (config: any, connections: any) => {
+  try {
+    let [selectedConnectionId] = await Promise.all([
+      localforage.getItem('selectedConnectionId'),
+    ]);
+
+    const update: { initialized: boolean; selectedConnectionId?: string } = {
+      initialized: true,
+    };
+
+    if (connections.length === 1) {
+      update.selectedConnectionId = connections[0].id;
+    } else {
+      const { defaultConnectionId } = config || {};
+      if (defaultConnectionId) {
+        const foundDefault = connections.find(
+          (c: any) => c.id === defaultConnectionId
+        );
+        if (Boolean(foundDefault)) {
+          update.selectedConnectionId = defaultConnectionId;
+        }
+      }
+
+      if (typeof selectedConnectionId === 'string') {
+        const selectedConnection = connections.find(
+          (c: any) => c.id === selectedConnectionId
+        );
+        if (Boolean(selectedConnection)) {
+          update.selectedConnectionId = selectedConnectionId;
+        }
+      }
+
+      const qs = queryString.parse(window.location.search);
+      const qsConnectionName = qs.connectionName;
+      if (qsConnectionName) {
+        const selectedConnection = connections.find(
+          (c: any) => c.name === qsConnectionName
+        );
+        if (Boolean(selectedConnection))
+          update.selectedConnectionId = selectedConnection.id;
+      }
+
+      const qsConnectionId = qs.connectionId;
+      if (qsConnectionId) {
+        const selectedConnection = connections.find(
+          (c: any) => c.id === qsConnectionId
+        );
+        if (Boolean(selectedConnection))
+          update.selectedConnectionId = selectedConnection.id;
+      }
+    }
+
+    useQueriesStore.setState(update);
+  } catch (error) {
+    console.error(error);
+    message.error('Error initializing application');
+  }
+};
 
 /**
  * Open a connection client for the currently selected connection if supported
